`SynthesisEngine`と`InferenceCore`を解体する (#686)
